move utility functions from periodic file to util file
